Backend/ems_project/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
companies==1.0.0
Django==6.0
django-cors-headers==4.2.0
django-filter==25.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.14.0
djangorestframework-stubs==3.16.6
drf-yasg==1.21.7
idna==3.11
inflection==0.5.1
packaging==25.0
phonenumbers==8.13.22
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
